Add Interface to DataSource classes
