+++ app/src/main/java/com/github/nakawai/newsreader/KoinInjector.kt
@@ -1,6 +1,8 @@
 package com.github.nakawai.newsreader
 
 import com.github.nakawai.newsreader.data.Repository
+import com.github.nakawai.newsreader.data.db.NYTimesLocalDataSourceImpl
+import com.github.nakawai.newsreader.data.network.NYTimesRemoteDataSourceImpl
 import com.github.nakawai.newsreader.domain.StoryAppService
 import com.github.nakawai.newsreader.ui.articles.ArticlesViewModel
 import com.github.nakawai.newsreader.ui.details.DetailsViewModel
@@ -9,7 +11,9 @@ import org.koin.android.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val appModule = module {
-    single { Repository() }
+    single { NYTimesRemoteDataSourceImpl() }
+    single { NYTimesLocalDataSourceImpl() }
+    single { Repository(get(), get()) }
     single { StoryAppService(get()) }
     viewModel { SectionsViewModel(get()) }
     viewModel { ArticlesViewModel(get()) }
